Rename the ANTLR package to grammar
