+++ grammar/generate.sh
@@ -5,4 +5,4 @@ rm ./*.tokens
 rm ./fhirpath*.go
 
 alias antlr4='java -Xmx500M -cp "./antlr-4.13.0-complete.jar:$CLASSPATH" org.antlr.v4.Tool'
-antlr4 -Dlanguage=Go -no-visitor -package antlrgen *.g4
+antlr4 -Dlanguage=Go -no-visitor -package grammar *.g4
